// types/theme.ts
export interface ThemeConfig {
  colors: {
    primary: {
      DEFAULT: string
      foreground: string
    }
    secondary: {
      DEFAULT: string
      foreground: string
    }
    card: {
      DEFAULT: string
      foreground: string
    }
    background: string
    foreground: string
    muted: {
      DEFAULT: string
      foreground: string
    }
    accent: {
      DEFAULT: string
      foreground: string
    }
    destructive: {
      DEFAULT: string
      foreground: string
    }
    success: {
      DEFAULT: string
      foreground: string
    }
  }
  borderRadius: {
    lg: string
    md: string
    sm: string
  }
  fontFamily: {
    sans: string
    heading: string
  }
}